% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks_secrets.R
\name{db_secrets_api}
\alias{db_secrets_api}
\title{Interact with the Databricks API for managing individual secrets.
The Databricks API allows for listing, creating,
and deleting secret scopes.}
\usage{
db_secrets_api(
  operation,
  workspace_url,
  access_token,
  scope_name,
  secret_name = NULL,
  secret_value = NULL,
  bytestring = FALSE
)
}
\arguments{
\item{operation}{The scopes operation to apply
("list", "put", or "delete")}

\item{workspace_url}{The url of the target databricks workspace}

\item{access_token}{The azure access token string which will be used for
authentication}

\item{scope_name}{The name of the scope containing the target secret}

\item{secret_name}{The name of the target secret}

\item{secret_value}{The value to assign to the target secret}

\item{bytestring}{Whether the target secret is storing a string of bytes or
characters}
}
\value{
The result of the API call
}
\description{
Interact with the Databricks API for managing individual secrets.
The Databricks API allows for listing, creating,
and deleting secret scopes.
}
\seealso{
Other Azure methods: 
\code{\link{az_authenticated_api_query}()},
\code{\link{db_secret_scopes_api}()},
\code{\link{store_databricks_access_token}()},
\code{\link{update_databricks_token}()}
}
\concept{Azure methods}
