% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_key_vault.R
\name{.init_key_vault}
\alias{.init_key_vault}
\title{Initialise an AzureKeyVault object for interacting with stored secrets. If
an AzureAuth access token object is not provided then the authentication
process is also initiated}
\usage{
.init_key_vault(vault_name, token_object)
}
\arguments{
\item{vault_name}{The name of the target Azure Key Vault}

\item{token_object}{A previously-generated AzureAuth token object
created for the Key Vault resource}
}
\value{
An AzureKeyVault::key_vault() object
}
\description{
Initialise an AzureKeyVault object for interacting with stored secrets. If
an AzureAuth access token object is not provided then the authentication
process is also initiated
}
\examples{
-
}
\seealso{
Other Key Vault methods: 
\code{\link{.secrets_operation}()},
\code{\link{kv_delete_secret}()},
\code{\link{kv_get_secret}()},
\code{\link{kv_list_secrets}()},
\code{\link{kv_write_secret}()}
}
\concept{Key Vault methods}
