% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_key_vault.R
\name{kv_delete_secret}
\alias{kv_delete_secret}
\title{Delete a specified secret from a given Key Vault}
\usage{
kv_delete_secret(vault_name, secret_name, confirm = TRUE, token_object = NULL)
}
\arguments{
\item{vault_name}{The name of the target Azure Key Vault}

\item{secret_name}{The name of the target secret to delete}

\item{confirm}{Whether to display an interactive prompt to confirm deletion
of stored secret}

\item{token_object}{(optional) A previously-generated AzureAuth token object
created for the Key Vault resource}
}
\description{
Delete a specified secret from a given Key Vault
}
\examples{
-
}
\seealso{
Other Key Vault methods: 
\code{\link{.init_key_vault}()},
\code{\link{.secrets_operation}()},
\code{\link{kv_get_secret}()},
\code{\link{kv_list_secrets}()},
\code{\link{kv_write_secret}()}
}
\concept{Key Vault methods}
