% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_key_vault.R
\name{kv_get_secret}
\alias{kv_get_secret}
\title{Retrieve the value of a specified secret from a given Key Vault}
\usage{
kv_get_secret(vault_name, secret_name, token_object = NULL)
}
\arguments{
\item{vault_name}{The name of the target Azure Key Vault}

\item{secret_name}{The name of the target secret to retrieve}

\item{token_object}{(optional) A previously-generated AzureAuth token object
created for the Key Vault resource}
}
\value{
The value of a secret stored in the given Key Vault
}
\description{
Retrieve the value of a specified secret from a given Key Vault
}
\examples{
-
}
\seealso{
Other Key Vault methods: 
\code{\link{.init_key_vault}()},
\code{\link{.secrets_operation}()},
\code{\link{kv_delete_secret}()},
\code{\link{kv_list_secrets}()},
\code{\link{kv_write_secret}()}
}
\concept{Key Vault methods}
