% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_key_vault.R
\name{kv_list_secrets}
\alias{kv_list_secrets}
\title{List the names of secrets stored in a given Azure Key Vault}
\usage{
kv_list_secrets(vault_name, token_object = NULL)
}
\arguments{
\item{vault_name}{The name of the target Azure Key Vault}

\item{token_object}{(optional) A previously-generated AzureAuth token object
created for the Key Vault resource}
}
\value{
A list of the (names of) secrets stored in the given Key Vault
}
\description{
List the names of secrets stored in a given Azure Key Vault
}
\seealso{
Other Key Vault methods: 
\code{\link{.init_key_vault}()},
\code{\link{.secrets_operation}()},
\code{\link{kv_delete_secret}()},
\code{\link{kv_get_secret}()},
\code{\link{kv_write_secret}()}
}
\concept{Key Vault methods}
