% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_key_vault.R
\name{kv_write_secret}
\alias{kv_write_secret}
\title{Create or overwrite a secret in an Azure Key Vault}
\usage{
kv_write_secret(
  vault_name,
  secret_name,
  secret_value,
  bytestring = FALSE,
  token_object = NULL
)
}
\arguments{
\item{vault_name}{The name of the target Azure Key Vault}

\item{secret_name}{The name of the target secret to update}

\item{secret_value}{The value to set the target secret to}

\item{bytestring}{Whether the secret value is a string of bytes, for storing
serialised objects}

\item{token_object}{(optional) A previously-generated AzureAuth token object
created for the Key Vault resource}
}
\description{
Create or overwrite a secret in an Azure Key Vault
}
\seealso{
Other Key Vault methods: 
\code{\link{.init_key_vault}()},
\code{\link{.secrets_operation}()},
\code{\link{kv_delete_secret}()},
\code{\link{kv_get_secret}()},
\code{\link{kv_list_secrets}()}
}
\concept{Key Vault methods}
