% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qiverse.azure-package.R
\docType{package}
\name{qiverse.azure-package}
\alias{qiverse.azure-package}
\alias{qiverse.azure}
\title{The qiverse.azure package.}
\description{
PowerBI, Sharepoint and Snowflake Connectivity
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AUS-DOH-Safety-and-Quality/qiverse/qiverse.azure}
  \item Report bugs at \url{https://github.com/AUS-DOH-Safety-and-Quality/qiverse/issues}
}

}
\author{
\strong{Maintainer}: Healthcare Quality Intelligence Unit (Western Australia Health) \email{safetyandquality@health.wa.gov.au}

Authors:
\itemize{
  \item Healthcare Quality Intelligence Unit (Western Australia Health) \email{safetyandquality@health.wa.gov.au}
}

}
